+++ src/yafowil/persistence.py
@@ -1,19 +1,15 @@
-from node.utils import UNSET
 from yafowil.base import factory
 
 
-factory.defaults['persist'] = False
 factory.doc['props']['persist'] = """\
 Marker that widget data should be considered in when using persistence
 mechanism.
 """
 
-factory.defaults['persist_target'] = UNSET
 factory.doc['props']['persist_target'] = """\
 Target when using persistence mechanism. If not set, target is widget name.
 """
 
-factory.defaults['persist_writer'] = UNSET
 factory.doc['props']['persist_writer'] = """\
 Callback used to persist data. Accepts model, target and value as parameters.
 """